bulk file api created and error handling has done

+++ src/main/java/com/currency/demo_app/dto/response/ExchangeRateResponseDTO.java
@@ -3,10 +3,12 @@ package com.currency.demo_app.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ExchangeRateResponseDTO {
     private String sourceCurrency;
     private String targetCurrency;
-    private double rate;
+    private BigDecimal rate;
 }
